Migrate scale gradient into a hook function
